requirements.txt
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
certifi==2015.11.20.1
kazoo==2.2.1
pyzmq==15.2.0
singledispatch==3.4.0.3
six==1.10.0
